+++ evmd/ante/evm_handler.go
@@ -18,6 +18,6 @@ func newMonoEVMAnteHandler(options ante.HandlerOptions) sdk.AnteHandler {
 		),
 		ante.NewTxListenerDecorator(options.PendingTxListener),
 	}
-	
+
 	return sdk.ChainAnteDecorators(decorators...)
 }