+++ evmd/app.go
@@ -2,6 +2,7 @@ package evmd
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -1143,6 +1144,22 @@ func (app *EVMD) GetTxConfig() client.TxConfig {
 
 func (app *EVMD) SetClientCtx(clientCtx client.Context) {
 	app.clientCtx = clientCtx
+}
+
+// Close unsubscribes from the CometBFT event bus (if set) and closes the underlying BaseApp.
+func (app *EVMD) Close() error {
+	var err error
+	if m, ok := app.GetMempool().(*evmmempool.ExperimentalEVMMempool); ok {
+		err = m.Close()
+	}
+	err = errors.Join(err, app.BaseApp.Close())
+	msg := "Application gracefully shutdown"
+	if err == nil {
+		app.Logger().Info(msg)
+	} else {
+		app.Logger().Error(msg, "error", err)
+	}
+	return err
 }
 
 // AutoCliOpts returns the autocli options for the app.